chore: refactor nhsbsa-doc functions
